requirements.txt
asgiref==3.5.2
autopep8==1.7.0
click==8.1.3
colorama==0.4.5
cssbeautifier==1.14.6
Django==3.2
django-taggit==1.2.0
djlint==1.19.1
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==5.0.0
jsbeautifier==1.14.6
Markdown==3.2.1
pathspec==0.10.1
pycodestyle==2.9.1
pytz==2022.5
PyYAML==6.0
regex==2022.9.13
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
zipp==3.9.0
